+++ gdal/src/test/scala/geotrellis/contrib/vlm/gdal/GDALErrorSpec.scala
@@ -10,7 +10,7 @@ import org.scalatest._
 class GDALErrorSpec extends FunSpec {
   val uri = Resource.path("img/c41078a1.tif")
   val token = GDALWarp.get_token(uri, Array())
-  val dataset: Int = GDALWarpOptions.EMPTY.datasetType
+  val dataset: GDALDataset.DatasetType = GDALWarpOptions.EMPTY.datasetType
 
   val sourceWindow: Array[Int] = Array(1000000, 1000000, 5000000, 5000000)
   val destWindow: Array[Int] = Array(500, 500)
@@ -18,13 +18,13 @@ class GDALErrorSpec extends FunSpec {
 
   describe("GDALErrors") {
     it("should return the ObjectNull error code") {
-      val result = GDALWarp.get_data(token, dataset, 1, sourceWindow, destWindow, 42, 1, buffer)
+      val result = GDALWarp.get_data(token, dataset.value, 1, sourceWindow, destWindow, 42, 1, buffer)
 
       assert(math.abs(result) == 10)
     }
 
     it("should return the IllegalArg error code") {
-      val result = GDALWarp.get_data(token, dataset, 1, sourceWindow, destWindow, 1, 1, buffer)
+      val result = GDALWarp.get_data(token, dataset.value, 1, sourceWindow, destWindow, 1, 1, buffer)
 
       assert(math.abs(result) == 5)
     }