+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Expose GDAL Metadata API, expose RasterSources Metadata API 
+
 ## [3.17.0] - 2019-07-23
 
 ### Added